Split and clean ApiHandler, add loopback bool param to configs
