+++ pyron-core/src/main/scala/com/cloudentity/pyron/domain/rule/rule.scala
@@ -17,6 +17,7 @@ case class RuleConf(
   criteria: EndpointMatchCriteria,
   target: TargetServiceRule,
   dropPathPrefix: Boolean,
+  loopback: Boolean,
   rewriteMethod: Option[RewriteMethod],
   rewritePath: Option[RewritePath],
   copyQueryOnRewrite: Option[Boolean],
