+++ pyron-engine/src/test/scala/com/cloudentity/pyron/rule/RuleBuilderSpec.scala
@@ -59,6 +59,7 @@ class RuleBuilderSpec extends WordSpec with MustMatchers with VertxSpec {
       ),
       target = StaticServiceRule(TargetHost(""), 9000, ssl = false),
       dropPathPrefix = true,
+      loopback = false,
       rewriteMethod = None,
       rewritePath = None,
       copyQueryOnRewrite = None,
